fix: Enforce `minItems` (#191)
* refactor: Add some whitespace between tests

* fix: Enforce `aerial-photo:run` minimum items

The `minItems` property was in the wrong place.

* fix: Enforce `aerial-photo:anomalies` minimum items

The `minItems` property was in the wrong place.

* fix: Enforce `film:id` minimum items

The `minItems` property was in the wrong place.

* fix: Enforce `film:physical_condition` minimum items

The `minItems` property was in the wrong place.

* fix: Enforce `film:physical_size` minimum items

The `minItems` property was in the wrong place.

* fix: Enforce `platform` minimum items

The `minItems` property was in the wrong place.

* fix: Enforce `mission` minimum items

The `minItems` property was in the wrong place.

* fix: Enforce `instruments` minimum items

The `minItems` property was in the wrong place.

* fix: Remove un-enforced `proj:epsg` minimum items

The `minItems` property was in the wrong place, but we don't actually
want to enforce it because `proj:epsg` can be `null`.